feat: Implement LyricsTab and MetadataTab components for improved song metadata and lyrics management; refactor VerificationDialog to utilize these new components
