+++ frontend/src/components/upload/VerificationDialog.tsx
@@ -11,14 +11,12 @@ import {
 import { Button } from "@/components/ui/button";
 import { Loader2, CheckCircle2 } from "lucide-react";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group";
-import { Label } from "@/components/ui/label";
-import { Card, CardContent } from "@/components/ui/card";
-import { ScrollArea } from "@/components/ui/scroll-area";
 import {
   LyricsOption,
   MetadataOption,
 } from "@/hooks/useYoutube";
+import { MetadataTab } from "./MetadataTab";
+import { LyricsTab } from "./LyricsTab";
 
 interface VerificationDialogProps {
   readonly isOpen: boolean;
@@ -69,41 +67,6 @@ export function VerificationDialog({
       selectedLyrics,
       selectedMetadata,
     });
-  };
-
-  // Function to render a preview of lyrics
-  const renderLyricsPreview = (lyrics?: string, isSynced = false) => {
-    if (!lyrics)
-      return (
-        <span className="text-muted-foreground italic">
-          No lyrics available
-        </span>
-      );
-
-    // Remove timestamps for synced lyrics preview
-    const cleanedLyrics = isSynced
-      ? lyrics
-          .split("\n")
-          .map((line) => line.replace(/^\[\d+:\d+\.\d+\]/g, "").trim())
-          .join("\n")
-      : lyrics;
-
-    // Only show first few lines
-    const lines = cleanedLyrics.split("\n").slice(0, 10);
-    return (
-      <>
-        {lines.map((line, i) => (
-          <p key={i} className={line.trim() === "" ? "h-4" : ""}>
-            {line}
-          </p>
-        ))}
-        {cleanedLyrics.split("\n").length > 10 && (
-          <p className="text-muted-foreground italic mt-2 border-t pt-1">
-            {cleanedLyrics.split("\n").length - 10} more lines (not shown)
-          </p>
-        )}
-      </>
-    );
   };
 
   return (
@@ -161,163 +124,21 @@ export function VerificationDialog({
             </TabsList>
             
             <TabsContent value="metadata" className="flex-1 overflow-hidden flex flex-col data-[state=inactive]:hidden">
-              <ScrollArea className="flex-1 pr-4">
-                {isLoadingMetadata ? (
-                  <div className="flex flex-col items-center justify-center py-8">
-                    <Loader2 className="h-8 w-8 animate-spin mb-4 text-primary" />
-                    <p className="text-center font-medium">Loading metadata options...</p>
-                    <p className="text-center text-sm text-muted-foreground mt-1">
-                      Searching music databases for the best match
-                    </p>
-                  </div>
-                ) : metadataOptions.length === 0 ? (
-                  <Card>
-                    <CardContent className="py-4">
-                      <p className="text-muted-foreground text-center">
-                        No additional metadata found
-                      </p>
-                      <p className="text-xs text-muted-foreground text-center mt-2">
-                        We'll use the information you provided in the previous
-                        step
-                      </p>
-                    </CardContent>
-                  </Card>
-                ) : (
-                  <RadioGroup 
-                    value={String(metadataOptions.indexOf(selectedMetadata || metadataOptions[0]))} 
-                    onValueChange={(value) => {
-                      const index = parseInt(value);
-                      setSelectedMetadata(metadataOptions[index] || null);
-                    }}
-                    className="space-y-4"
-                  >
-                    {metadataOptions.map((option, index) => (
-                      <div key={option.id || index} className="flex items-start space-x-2">
-                        <RadioGroupItem value={String(index)} id={`metadata-${index}`} className="mt-1" />
-                        <div className="flex-1">
-                          <Label 
-                            htmlFor={`metadata-${index}`} 
-                            className="flex flex-col space-y-1 cursor-pointer"
-                          >
-                            <Card className={`hover:border-primary ${selectedMetadata === option ? 'border-primary bg-primary/5' : ''}`}>
-                              <CardContent className="p-3">
-                                <div className="flex justify-between items-start">
-                                  <div className="flex-1">
-                                    <h4 className="font-medium text-sm">{option.title}</h4>
-                                    <p className="text-sm text-muted-foreground">
-                                      {option.artist}
-                                      {option.album && ` • ${option.album}`}
-                                    </p>
-                                    {(option.year || option.genre || option.language) && (
-                                      <div className="mt-1 flex flex-wrap gap-1">
-                                        {option.year && (
-                                          <span className="px-2 py-0.5 text-xs bg-secondary rounded-full">
-                                            {option.year}
-                                          </span>
-                                        )}
-                                        {option.genre && (
-                                          <span className="px-2 py-0.5 text-xs bg-secondary rounded-full">
-                                            {option.genre}
-                                          </span>
-                                        )}
-                                        {option.language && (
-                                          <span className="px-2 py-0.5 text-xs bg-secondary rounded-full">
-                                            {option.language}
-                                          </span>
-                                        )}
-                                      </div>
-                                    )}
-                                  </div>
-                                  {option.source && (
-                                    <span className="text-xs bg-muted px-2 py-1 rounded text-muted-foreground">
-                                      {option.source}
-                                    </span>
-                                  )}
-                                </div>
-                              </CardContent>
-                            </Card>
-                          </Label>
-                        </div>
-                      </div>
-                    ))}
-                  </RadioGroup>
-                )}
-              </ScrollArea>
+              <MetadataTab 
+                options={metadataOptions}
+                selectedOption={selectedMetadata}
+                onSelectionChange={setSelectedMetadata}
+                isLoading={isLoadingMetadata}
+              />
             </TabsContent>
 
             <TabsContent value="lyrics" className="flex-1 overflow-hidden flex flex-col data-[state=inactive]:hidden">
-              <ScrollArea className="flex-1 pr-4">
-                {isLoadingLyrics ? (
-                  <div className="flex flex-col items-center justify-center py-8">
-                    <Loader2 className="h-8 w-8 animate-spin mb-4 text-primary" />
-                    <p className="text-center font-medium">Loading lyrics options...</p>
-                    <p className="text-center text-sm text-muted-foreground mt-1">
-                      Searching for lyrics that match your song
-                    </p>
-                  </div>
-                ) : lyricsOptions.length === 0 ? (
-                  <Card>
-                    <CardContent className="py-4">
-                      <p className="text-muted-foreground text-center">
-                        No lyrics found
-                      </p>
-                      <p className="text-xs text-muted-foreground text-center mt-2">
-                        You can add lyrics manually later
-                      </p>
-                    </CardContent>
-                  </Card>
-                ) : (
-                  <RadioGroup 
-                    value={String(lyricsOptions.indexOf(selectedLyrics || lyricsOptions[0]))} 
-                    onValueChange={(value) => {
-                      const index = parseInt(value);
-                      setSelectedLyrics(lyricsOptions[index] || null);
-                    }}
-                    className="space-y-4"
-                  >
-                    {lyricsOptions.map((option, index) => (
-                      <div key={option.id || index} className="flex items-start space-x-2">
-                        <RadioGroupItem value={String(index)} id={`lyrics-${index}`} className="mt-1" />
-                        <div className="flex-1">
-                          <Label 
-                            htmlFor={`lyrics-${index}`} 
-                            className="flex flex-col space-y-1 cursor-pointer"
-                          >
-                            <Card className={`hover:border-primary ${selectedLyrics === option ? 'border-primary bg-primary/5' : ''}`}>
-                              <CardContent className="p-3">
-                                <div className="flex justify-between items-start mb-2">
-                                  <div className="flex items-center gap-2">
-                                    <h4 className="font-medium text-sm">
-                                      {option.trackName || "Lyrics"}
-                                    </h4>
-                                    {option.syncedLyrics && (
-                                      <span className="px-2 py-0.5 text-xs bg-green-100 dark:bg-green-900 text-green-800 dark:text-green-200 rounded-full">
-                                        Synchronized
-                                      </span>
-                                    )}
-                                  </div>
-                                  {option.source && (
-                                    <span className="text-xs bg-muted px-2 py-1 rounded text-muted-foreground">
-                                      {option.source}
-                                    </span>
-                                  )}
-                                </div>
-                                
-                                <div className="text-sm mt-2 border-t pt-2 text-muted-foreground">
-                                  {renderLyricsPreview(
-                                    option.syncedLyrics || option.plainLyrics, 
-                                    !!option.syncedLyrics
-                                  )}
-                                </div>
-                              </CardContent>
-                            </Card>
-                          </Label>
-                        </div>
-                      </div>
-                    ))}
-                  </RadioGroup>
-                )}
-              </ScrollArea>
+              <LyricsTab 
+                options={lyricsOptions}
+                selectedOption={selectedLyrics}
+                onSelectionChange={setSelectedLyrics}
+                isLoading={isLoadingLyrics}
+              />
             </TabsContent>
           </Tabs>
         )}
